i36: Split AppBar component and remove Menu require cycle
