+++ src/library/components/HomeBar.js
@@ -1,12 +1,8 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import { View, Image, TouchableOpacity, Text } from 'react-native';
-
+import { View, Image, TouchableOpacity } from 'react-native';
 import { create } from 'library/utils/normalize.js'
 import R from 'res/R';
-import { IconButton, Menu, MenuDivider, MenuItem } from 'library/components';
-
-// HomeBar - Used in the Home Screen
+import { Menu, MenuDivider, MenuItem } from './Menu';
 
 class HomeBar extends React.Component {
   state = { showingMenu: false };
@@ -50,64 +46,16 @@ class HomeBar extends React.Component {
   }
 }
 
-// NavBar - Used for navigation
-
-class NavBar extends React.Component {
-  static propTypes = {
-    label: PropTypes.string.isRequired,
-  };
-
-  static defaultProps = {
-    label: 'Screen Label',
-  };
-
-  render() {
-    const {
-      label,
-      onClose,
-      children,
-    } = this.props;
-    return (
-      <View style={styles.navContainer}>
-        <View style={[styles.itemsContainer, { flexDirection: 'row' }]}>
-          <IconButton icon='close' color={R.COLORS.graniteGray} onPress={onClose}/>
-          <Text style={styles.navLabel}>{label}</Text>
-        </View>
-        <View style={[styles.itemsContainer, { flexDirection: 'row-reverse' }]}>
-          {children}
-        </View>
-      </View>
-    );
-  }
-}
-
-// StyleSheet
-
 const styles = create({
-
   homeContainer: {
     flexDirection: 'row',
     height: 72,
     paddingHorizontal: 16,
   },
 
-  navContainer: {
-    flexDirection: 'row',
-    height: 56,
-    paddingHorizontal: 4,
-    backgroundColor: R.COLORS.justWhite,
-  },
-
   itemsContainer: {
     flex: 1,
     alignItems: 'center',
-  },
-
-  navLabel: {
-    color: R.COLORS.asphaltGray,
-    left: 12,
-    top: 1,
-    ...R.TYPOGRAPHY.headline5,
   },
 
   accountButton: {
@@ -127,10 +75,4 @@ const styles = create({
   },
 });
 
-// Export
-
-const AppBar = (props) => {
-  const Component = props.type === 'home' ? HomeBar : NavBar;
-  return <Component {...props} />
-}
-export default AppBar;
+export default HomeBar;
