+++ src/library/components/Menu/index.js
@@ -1,0 +1,3 @@
+export { default as Menu } from './Menu'
+export { default as MenuDivider } from './MenuDivider'
+export { default as MenuItem } from './MenuItem'
